Resolution of Internal server error bug

+++ src/main/java/org/traccar/api/resource/ClientPermissionResource.java
@@ -6,10 +6,15 @@ import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import org.traccar.api.BaseResource;
 import org.traccar.api.security.PermissionsService;
+import org.traccar.model.ClientDevice;
 import org.traccar.model.LinkType;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.traccar.storage.StorageException;
+import org.traccar.storage.query.Columns;
+import org.traccar.storage.query.Condition;
+import org.traccar.storage.query.Request;
 
 
 @Path("client/permissions")
@@ -25,15 +30,37 @@ public class ClientPermissionResource extends BaseResource {
     @POST
     public Response linkClientDevice(@QueryParam("clientId") long clientId, @QueryParam("deviceId") long deviceId) throws Exception{
         //LOGGER.info("Received POST request -> clientId: {}, deviceId: {}", clientId, deviceId);
-        permissionsService.link(LinkType.CLIENT_DEVICE, clientId, deviceId);
-        return Response.ok("{\"status\":\"Linked success\"}").build();
+
+        if(!validateClientDeviceLink(clientId,deviceId)){
+            permissionsService.link(LinkType.CLIENT_DEVICE, clientId, deviceId);
+            return Response.ok("{\"status\":\"Linked successfully.\"}").build();
+        }else{
+            return Response.ok("{\"status\":\"Device already linked to the client.\"}").build();
+        }
+
     }
 
     @DELETE
     public Response unlinkClientDevice(@QueryParam("clientId") long clientId, @QueryParam("deviceId") long deviceId) throws Exception{
-        permissionsService.unlink(LinkType.CLIENT_DEVICE, clientId, deviceId);
-        //return Response.ok().build();
-        return Response.ok("{\"status\":\"Unlinked success\"}").build();
+        if(validateClientDeviceLink(clientId,deviceId)){
+            permissionsService.unlink(LinkType.CLIENT_DEVICE, clientId, deviceId);
+            return Response.ok("{\"status\":\"Unlinked success\"}").build();
+        }else{
+            return Response.ok("{\"status\":\"The Client-device link doesnot exist.\"}").build();
+        }
+    }
+
+    public boolean validateClientDeviceLink(long clientId, long deviceId) throws StorageException {
+        // Query the database for a record matching both groupId and deviceId
+        ClientDevice link = storage.getObject(ClientDevice.class, new Request(
+                new Columns.All(),
+                new Condition.And(
+                        new Condition.Equals("clientid", clientId),
+                        new Condition.Equals("deviceid", deviceId)
+                )));
+
+        // If the record exists, return true; otherwise, false
+        return link != null;
     }
 
 }
